Update telegram-chat-parser.py
The output file is now opened with the encoding "utf-8-sig" to support Hindi characters.
In this version, the code prompts the user to enter the output file path and name. The input() function is used to get the input from the user. The newline parameter is also added to the open() function for proper handling of newlines in the CSV file.